requirements.txt
flask
gunicorn
torch==2.1.0+cpu
torchvision==0.16.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
timm==0.9.12
xgboost==2.0.3
scikit-learn==1.3.2
numpy==1.26.2
Pillow==10.1.0